\name{gapfillNcdfCoreprocess}
\alias{gapfillNcdfCoreprocess}
\title{helper function for gapfillNcdf}
\description{####################### gapfill function for single core ####################}
\usage{gapfillNcdfCoreprocess(args.call.SSA, datacube, datapts.n, dims.cycle.id, 
    dims.cycle.length, dims.process.id, dims.process.length, 
    file.name, first.guess, ind.process.cube, iter.gridind, iter.nr, 
    iters.n, MSSA, print.status, reproducible)}
\arguments{
  \item{args.call.SSA}{
}
  \item{datacube}{
}
  \item{datapts.n}{
}
  \item{dims.cycle.id}{
}
  \item{dims.cycle.length}{
}
  \item{dims.process.id}{
}
  \item{dims.process.length}{
}
  \item{file.name}{
}
  \item{first.guess}{
}
  \item{ind.process.cube}{
}
  \item{iter.gridind}{
}
  \item{iter.nr}{
}
  \item{iters.n}{
}
  \item{MSSA}{
}
  \item{print.status}{
}
  \item{reproducible}{
}
}
\details{helper function for gapfillNcdf performs each individual series/grid 
extracion and handing it over to gapfillSSA.}


\author{Jannis v. Buttlar}





