\name{plotDecomposition}
\alias{plotDecomposition}
\title{Plot the results of a SSA decomposition}
\description{This function visualizes the results from a call to filterTSeriesSSA}
\usage{plotDecomposition(decomp.results, time.vector = c(), n.magnif = 3)}\arguments{
  \item{decomp.results}{list: object to plot: results from a call to filterTSeriesSSA}
  \item{time.vector}{R Date/time object: optional time vector used to label x axes.}
  \item{n.magnif}{amount of positions where to plot the different magnifications.
This results in the amount of columns in the plot}
}

\value{Nothing is returned.
do plots}

\author{Jannis v. Buttlar}





