\name{rbindMod}
\alias{rbindMod}
\title{helper function for gapfillNcdf}
\description{###############  combine data from foreach iteration ########################}
\usage{rbindMod(...)}\arguments{
  \item{\dots}{
}
}
\details{helper function for gapfillNcdf that combines the foreach output
in a convenient way.}


\author{Jannis v. Buttlar}





