\name{spectral.methods-package}
\alias{spectral.methods-package}
\alias{spectral.methods}
\docType{package}
\title{Collection of SSA tools for spectral analysis}
\description{The package contains some implementations of SSA for the gap filling (GapfillSSA() ) and spectral decomposition (FilterTseriesSSA() ) of time series. It uses the SSA algorithm developed by Korobeneykov (2009) and implemented in Rssa which has to be downloaded from Cran or http://github.com/asl/rssa.}
\details{
\tabular{ll}{Package: \tab spectral.methods\cr
Title: \tab Collection of SSA tools for spectral analysis\cr
Version: \tab 0.7.0.115\cr
Date: \tab 2013-08-15\cr
Author: \tab Jannis v. Buttlar\cr
Maintainer: \tab Jannis v. Buttlar <jbuttlar@bgc-jena.mpg.de>\cr
Imports: \tab Rssa, raster, nnet, abind, RNetCDF, ncdf.tools, foreach, doMC, JBTools, snow, snowfall, DistributionUtils, RColorBrewer\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
Depends: \tab R (>= 2.10.0)\cr}
}
\author{Jannis v. Buttlar}

\keyword{ package }


