\name{spectral.methods-package}
\alias{spectral.methods-package}
\alias{spectral.methods}
\docType{package}
\title{Singular Spectrum Analysis (SSA) tools for time series analysis}
\description{This package contains some implementations of Singular Spectrum Analysis (SSA) for the gapfilling (gapfillSSA() ) and spectral decomposition (filterTseriesSSA() ) of time series. It contains the code used by Buttlar et. al. (2014), Nonlinear Processes in Geophysics (gapfillNcdf). In addition, the iterative SSA gapfilling of Kondrashov and Ghil (2006) is implemented. All SSA calculations are done via the truncated and fast SSA algorithm of Korobeneykov (2010) (package Rssa).}
\details{
\tabular{ll}{Package: \tab spectral.methods\cr
Title: \tab Singular Spectrum Analysis (SSA) tools for time series analysis\cr
Version: \tab 0.7.2.123\cr
Date: \tab 2014-01-01\cr
Author: \tab Jannis v. Buttlar\cr
Maintainer: \tab Jannis v. Buttlar <jbuttlar@bgc-jena.mpg.de>\cr
Imports: \tab Rssa, raster, nnet, abind, RNetCDF, ncdf.tools, foreach, JBTools, parallel, DistributionUtils, RColorBrewer\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
Depends: \tab R (>= 2.10.0)\cr}
}
\author{Jannis v. Buttlar}

\keyword{ package }


