\name{spectral.methods-package}
\alias{spectral.methods-package}
\alias{spectral.methods}
\docType{package}
\title{Singular Spectrum Analysis (SSA) Tools for Time Series Analysis}
\description{Contains some implementations of Singular Spectrum Analysis (SSA) for the gapfilling and spectral decomposition of time series. It contains the code used by Buttlar et. al. (2014), Nonlinear Processes in Geophysics. In addition, the iterative SSA gapfilling method of Kondrashov and Ghil (2006) is implemented. All SSA calculations are done via the truncated and fast SSA algorithm of Korobeynikov (2010) (package 'Rssa').}
\details{
\tabular{ll}{Package: \tab spectral.methods\cr
Title: \tab Singular Spectrum Analysis (SSA) Tools for Time Series Analysis\cr
Version: \tab 0.7.2.133\cr
Date: \tab 2015-05-20\cr
Author: \tab Jannis v. Buttlar\cr
Maintainer: \tab Jannis v. Buttlar <jbuttlar@bgc-jena.mpg.de>\cr
Imports: \tab Rssa (>= 0.11), raster, nnet, abind, RNetCDF, ncdf.tools, foreach, JBTools, DistributionUtils, RColorBrewer\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
Depends: \tab R (>= 3.0.0)\cr}
}
\author{Jannis v. Buttlar}

\keyword{ package }


